movieland-service-default/src/main/java/com/khrushch/movieland/service/DefaultReviewService.java
package com.khrushch.movieland.service;

import com.khrushch.movieland.dao.ReviewDao;
import com.khrushch.movieland.dto.ReviewDto;
import com.khrushch.movieland.model.Movie;
import com.khrushch.movieland.model.Review;
import com.khrushch.movieland.model.User;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class DefaultReviewService implements ReviewService {
    private ReviewDao reviewDao;

    private SecurityService securityService;

    @Override
    public void addReview(ReviewDto dto) {
        User user = securityService.getUserByUuid(dto.getUuid());

        Review review = new Review();
        review.setUser(user);
        review.setMovieId(dto.getMovieId());
        review.setText(dto.getText());

        reviewDao.addReview(review);
    }

    @Override
    public void enrich(Movie movie) {
        List<Review> reviews = reviewDao.getByMovieId(movie.getId());
        movie.setReviews(reviews);
    }

    @Autowired
    public void setReviewDao(ReviewDao reviewDao) {
        this.reviewDao = reviewDao;
    }

    @Autowired
    public void setSecurityService(SecurityService securityService) {
        this.securityService = securityService;
    }
}
